+++ gcm-simd.cpp
@@ -24,16 +24,6 @@
 # undef CRYPTOPP_ARM_PMULL_AVAILABLE
 #endif
 
-#if defined(__linux__)
-# include <sys/auxv.h>
-# ifndef HWCAP_PMULL
-# define HWCAP_PMULL (1 << 4)
-# endif
-# ifndef HWCAP2_PMULL
-# define HWCAP2_PMULL (1 << 1)
-# endif
-#endif
-
 #if (CRYPTOPP_CLMUL_AVAILABLE)
 # include "tmmintrin.h"
 # include "wmmintrin.h"
@@ -203,7 +193,7 @@ extern "C" {
 #endif  // Not CRYPTOPP_MS_STYLE_INLINE_ASSEMBLY
 
 #if (CRYPTOPP_BOOL_ARM32 || CRYPTOPP_BOOL_ARM64)
-bool CPU_TryPMULL_ARMV8()
+bool CPU_ProbePMULL()
 {
 #if (CRYPTOPP_ARM_PMULL_AVAILABLE)
 # if defined(CRYPTOPP_MS_STYLE_INLINE_ASSEMBLY)
@@ -230,17 +220,6 @@ bool CPU_TryPMULL_ARMV8()
     }
     return result;
 # else
-#   if defined(__ANDROID__) && (defined(__aarch64__) || defined(__aarch32__))
-    if (android_getCpuFeatures() & ANDROID_CPU_ARM64_FEATURE_PMULL)
-        return true;
-    // https://sourceware.org/ml/libc-help/2017-08/msg00012.html
-#   elif defined(__linux__) && defined(__aarch64__)
-    if (getauxval(AT_HWCAP) & HWCAP_PMULL)
-        return true;
-#   elif defined(__linux__) && defined(__aarch32__)
-    if (getauxval(AT_HWCAP2) & HWCAP2_PMULL)
-        return true;
-#   endif
 
     // longjmp and clobber warnings. Volatile is required.
     // http://github.com/weidai11/cryptopp/issues/24 and http://stackoverflow.com/q/7721854
